projects-orchestrator/requirements.txt
pyyaml
tornado>=5.1
requests
docker
